feat: half-star ratings, TSG-style charts, always-visible empty states
- Rating system: int→double (0.5-5 half-star increments), DB migration, validation
- Chart colors: vibrant multi-color bars/pies (categorical palette + uniform blue)
- Card layout: bordered panes with dividers matching TheStoryGraph
- Empty states: charts always render (empty circles for pies, "no data" for bars)
- StarRating: half-star click zones with SVG clipPath
- RatingChart: half-star labels (★½, ★★, etc.)
- i18n: stats.noData key (en/uk)

+++ backend/src/Domain/Entities/UserRating.cs
@@ -6,7 +6,7 @@ public class UserRating
     public Guid UserId { get; set; }
     public Guid SiteId { get; set; }
     public Guid EditionId { get; set; }
-    public int Rating { get; set; } // 1-5
+    public double Rating { get; set; } // 0.5-5 (half-star increments)
     public string? ReviewText { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
     public DateTimeOffset UpdatedAt { get; set; }
